Extract backported code to its own file
